Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 
 int main(){
-    int a,b;
+    double a,b;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%lf %lf %c",&a,&b,&c);
     if(c == '+'){
-        printf("%d",a+b);
+        printf("%lf",a+b);
     }else if(c == '*'){
-        printf("%d",a*b);
+        printf("%lf",a*b);
     }else if(c == '/'){
-        printf("%d",a/b);
+        printf("%lf",a/b);
     }else if(c == '-'){
-        printf("%d",a-b);
+        printf("%lf",a-b);
     }
     return 0;
 }
